feat: LHDN MyInvois compliance — entities, migrations, tenancy, bug fixes
- BUG 1: InvoiceSequence table + FOR UPDATE pessimistic lock replaces racy MAX() scan
- BUG 2: PromoteAsync guard rejects already-promoted invoices
- BUG 3: GetByIdAsync now filters IsActive
- BUG 4/5: Unique index on InvoiceNo; VatPercentage.Percentage changed to decimal
- BUG 6: Promote endpoint switched to [FromBody] SalesInvoicePromoteDto
- C9: Overpayment guard in SalesPaymentService
- TENANT 1: DI throws instead of silently falling back to master DB
- TENANT 2: SystemSettingsSeeder called during new tenant provisioning
- TENANT 3: IsDeactivated flag on ITenantContext + TenantGuardMiddleware (403)
- TENANT 4: Logo uploads namespaced by tenantId
- TENANT 6: TenantService.GetAsync filters IsActive
- Phase 1: TaxCode entity (codes 01–07) + unique filtered index + 7 LHDN seed rows
- Phase 2: CompanyInfo LHDN fields (TIN, MSICCode, MyInvois credentials); Customer buyer identity (TIN, RegistrationType, RegistrationNumber)
- Phase 3: SalesInvoice e-invoice fields (UUID, LongId, Status, TypeCode, dates, buyer snapshot, FxRate); SalesInvoiceLine UBL fields (ClassificationCode, TaxTypeCode, UnitCode); ReturnLog.OriginalEInvoiceUUID; buyer snapshot set at promotion time
- Phase 4: EInvoiceSubmission + EInvoiceValidationError audit trail entities
- Phase 5: ReportTemplate (master DB) + TenantReportTemplate (tenant DB) two-tier template cascade
- Seeder: Tax Payable GL account (22000, Liability) added; LHDN TaxCode rows seeded per new tenant
- Migrations: BugFixes_InvoiceSequence_InvoiceNoIndex_VatDecimal, Compliance_TaxCodes_EInvoice_ReportTemplates (tenant), Compliance_ReportTemplates (master)

+++ src/LitXusCount.Application/Settings/VatPercentages/VatPercentageDto.cs
@@ -1,5 +1,5 @@
 namespace LitXusCount.Application.Settings.VatPercentages;
 
-public record VatPercentageDto(long Id, string Name, double Percentage, bool IsDefault, bool IsActive);
+public record VatPercentageDto(long Id, string Name, decimal Percentage, bool IsDefault, bool IsActive);
 
-public record VatPercentageUpsertDto(string Name, double Percentage);
+public record VatPercentageUpsertDto(string Name, decimal Percentage);
